Pavels' patches
