+++ src/android/com/adobe/phonegap/push/IncomingCallActivity.kt
@@ -1,142 +1,221 @@
 package com.adobe.phonegap.push
 
+import android.Manifest
+import android.annotation.SuppressLint
 import android.app.Activity
+import android.app.KeyguardManager
+import android.content.BroadcastReceiver
+import android.content.Context
+import android.content.Intent
+import android.content.IntentFilter
+import android.content.pm.PackageManager
+import android.graphics.drawable.Drawable
+import android.os.Build
+import android.os.Bundle
+import android.os.Handler
+import android.os.Looper
+import android.util.Log
+import android.view.WindowManager
+import android.widget.Button
+import android.widget.ImageView
+import android.widget.TextView
+import androidx.core.app.ActivityCompat
+import androidx.core.app.NotificationCompat
+import androidx.core.app.NotificationManagerCompat
+import androidx.vectordrawable.graphics.drawable.Animatable2Compat
+import androidx.vectordrawable.graphics.drawable.AnimatedVectorDrawableCompat
+import com.trusted.care.staging.R
+
+private const val POST_NOTIFICATIONS_REQUEST_CODE = 8234
 
 class IncomingCallActivity : Activity() {
-  var caller = ""
-  @Override
-  protected fun onCreate(savedInstanceState: Bundle?) {
-    super.onCreate(savedInstanceState)
-    setContentView(getResources().getIdentifier("activity_incoming_call", "layout", getPackageName()))
-    getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN, WindowManager.LayoutParams.FLAG_FULLSCREEN)
-    getWindow().addFlags(WindowManager.LayoutParams.FLAG_SHOW_WHEN_LOCKED)
-    getWindow().addFlags(WindowManager.LayoutParams.FLAG_DISMISS_KEYGUARD)
-    getWindow().addFlags(WindowManager.LayoutParams.FLAG_TURN_SCREEN_ON)
-    getWindow().addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
-    instance = this
-    caller = getIntent().getExtras().getString("caller")
-    (findViewById(getResources().getIdentifier("tvCaller", "id", getPackageName())) as TextView).setText(caller)
-    val btnAccept: Button = findViewById(getResources().getIdentifier("btnAccept", "id", getPackageName()))
-    val btnDecline: Button = findViewById(getResources().getIdentifier("btnDecline", "id", getPackageName()))
-    btnAccept.setOnClickListener { v -> requestPhoneUnlock() }
-    btnDecline.setOnClickListener { v -> declineIncomingVoIP() }
-    val animatedCircle: ImageView = findViewById(getResources().getIdentifier("ivAnimatedCircle", "id", getPackageName()))
-    val drawableCompat: AnimatedVectorDrawableCompat = AnimatedVectorDrawableCompat.create(this, getResources().getIdentifier("circle_animation_avd", "drawable", getPackageName()))
-    animatedCircle.setImageDrawable(drawableCompat)
-    drawableCompat.registerAnimationCallback(object : AnimationCallback() {
-      @NonNull
-      private val fHandler: Handler = Handler(Looper.getMainLooper())
-      @Override
-      fun onAnimationEnd(drawable: Drawable?) {
-        super.onAnimationEnd(drawable)
-        if (instance != null) {
-          fHandler.post(drawableCompat::start)
-        }
-      }
-    })
-    drawableCompat.start()
-  }
 
-  @Override
-  fun onBackPressed() {
-    // Do nothing on back button
-  }
+    var caller: String = ""
 
-  fun requestPhoneUnlock() {
-    val km: KeyguardManager = getSystemService(Context.KEYGUARD_SERVICE) as KeyguardManager
-    if (km.isKeyguardLocked()) {
-      if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-        km.requestDismissKeyguard(this, object : KeyguardDismissCallback() {
-          @Override
-          fun onDismissSucceeded() {
-            super.onDismissSucceeded()
-            acceptIncomingVoIP()
-          }
+    override fun onCreate(savedInstanceState: Bundle?) {
+        showWhenLockedAndTurnScreenOn()
+        super.onCreate(savedInstanceState)
+        Log.d("", "IncomingCallActivity.onCreate()")
+        setContentView(R.layout.activity_incoming_call)
 
-          @Override
-          fun onDismissCancelled() {
-            super.onDismissCancelled()
-          }
+        window.setFlags(
+            WindowManager.LayoutParams.FLAG_FULLSCREEN,
+            WindowManager.LayoutParams.FLAG_FULLSCREEN
+        )
+        //window.addFlags(WindowManager.LayoutParams.FLAG_SHOW_WHEN_LOCKED)
+        //window.addFlags(WindowManager.LayoutParams.FLAG_DISMISS_KEYGUARD)
+        //window.addFlags(WindowManager.LayoutParams.FLAG_TURN_SCREEN_ON)
+        //window.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
+        instance = this
 
-          @Override
-          fun onDismissError() {
-            super.onDismissError()
-          }
+        caller = intent?.extras?.getString("caller") ?: ""
+        (findViewById<TextView>(R.id.tvCaller)).text = caller
+        val btnAccept: Button = findViewById(R.id.btnAccept)
+        val btnDecline: Button = findViewById(R.id.btnDecline)
+        btnAccept.setOnClickListener { v -> requestPhoneUnlock() }
+        btnDecline.setOnClickListener { v -> declineIncomingVoIP() }
+
+        val animatedCircle: ImageView = findViewById(R.id.ivAnimatedCircle)
+        val drawableCompat = AnimatedVectorDrawableCompat.create(this, R.drawable.circle_animation_avd)
+        animatedCircle.setImageDrawable(drawableCompat)
+        drawableCompat?.registerAnimationCallback(object : Animatable2Compat.AnimationCallback() {
+            private val fHandler = Handler(Looper.getMainLooper())
+            override fun onAnimationEnd(drawable: Drawable?) {
+                super.onAnimationEnd(drawable)
+                if (instance != null) {
+                    fHandler.post(drawableCompat::start)
+                }
+            }
         })
-      } else {
-        acceptIncomingVoIP()
-        if (km.isKeyguardSecure()) {
-          // Register receiver for dismissing "Unlock Screen" notification
-          phoneUnlockBR = PhoneUnlockBroadcastReceiver()
-          val filter = IntentFilter()
-          filter.addAction(Intent.ACTION_USER_PRESENT)
-          this.getApplicationContext().registerReceiver(phoneUnlockBR, filter)
-          showUnlockScreenNotification()
+        drawableCompat?.start()
+    }
+
+    private fun showWhenLockedAndTurnScreenOn() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O_MR1) {
+            setShowWhenLocked(true)
+            setTurnScreenOn(true)
         } else {
-          val myLock: KeyguardManager.KeyguardLock = km.newKeyguardLock("AnswerCall")
-          myLock.disableKeyguard()
+            window.addFlags(
+                WindowManager.LayoutParams.FLAG_SHOW_WHEN_LOCKED
+                        or WindowManager.LayoutParams.FLAG_TURN_SCREEN_ON
+            )
         }
-      }
-    } else {
-      acceptIncomingVoIP()
     }
-  }
 
-  fun acceptIncomingVoIP() {
-    val acceptIntent = Intent(VOIP_ACCEPT)
-    sendBroadcast(acceptIntent)
-  }
-
-  fun declineIncomingVoIP() {
-    val declineIntent = Intent(VOIP_DECLINE)
-    sendBroadcast(declineIntent)
-  }
-
-  private fun showUnlockScreenNotification() {
-    val notificationBuilder: NotificationCompat.Builder = Builder(this, PushConstants.DEFAULT_CHANNEL_ID)
-        .setSmallIcon(getResources().getIdentifier("pushicon", "drawable", getPackageName()))
-        .setContentTitle("Ongoing call with $caller")
-        .setContentText("Please unlock your device to continue")
-        .setPriority(NotificationCompat.PRIORITY_HIGH)
-        .setCategory(NotificationCompat.CATEGORY_MESSAGE)
-        .setAutoCancel(false)
-        .setOngoing(true)
-        .setStyle(BigTextStyle())
-        .setSound(null)
-    val ongoingCallNotification: Notification = notificationBuilder.build()
-    val notificationManager: NotificationManagerCompat = NotificationManagerCompat.from(this.getApplicationContext())
-    // Display notification
-    notificationManager.notify(NOTIFICATION_MESSAGE_ID, ongoingCallNotification)
-  }
-
-  @Override
-  protected fun onDestroy() {
-    super.onDestroy()
-    instance = null
-  }
-
-  class PhoneUnlockBroadcastReceiver : BroadcastReceiver() {
-    @Override
-    fun onReceive(context: Context, intent: Intent) {
-      if (intent.getAction().equals(Intent.ACTION_USER_PRESENT)) {
-        dismissUnlockScreenNotification(context.getApplicationContext())
-      }
+    override fun onBackPressed() {
+        // Do nothing on back button
     }
-  }
 
-  companion object {
-    const val VOIP_CONNECTED = "connected"
-    const val VOIP_ACCEPT = "pickup"
-    const val VOIP_DECLINE = "declined_callee"
-    private const val NOTIFICATION_MESSAGE_ID = 1337
-    var instance: IncomingCallActivity? = null
-    var phoneUnlockBR: PhoneUnlockBroadcastReceiver? = null
-    fun dismissUnlockScreenNotification(applicationContext: Context) {
-      NotificationManagerCompat.from(applicationContext).cancel(NOTIFICATION_MESSAGE_ID)
-      if (phoneUnlockBR != null) {
-        applicationContext.unregisterReceiver(phoneUnlockBR)
-        phoneUnlockBR = null
-      }
+    private fun requestPhoneUnlock() {
+        val km = getSystemService(Context.KEYGUARD_SERVICE) as KeyguardManager
+        val context = this.applicationContext
+        if (km.isKeyguardLocked) {
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+                km.requestDismissKeyguard(this, object : KeyguardManager.KeyguardDismissCallback() {
+                    override fun onDismissSucceeded() {
+                        super.onDismissSucceeded()
+                        acceptIncomingVoIP()
+                    }
+
+                    override fun onDismissCancelled() {
+                        super.onDismissCancelled()
+                    }
+
+                    override fun onDismissError() {
+                        super.onDismissError()
+                    }
+                })
+            } else {
+                acceptIncomingVoIP()
+                if (km.isKeyguardSecure) {
+                    // Register receiver for dismissing "Unlock Screen" notification
+                    phoneUnlockBR = PhoneUnlockBroadcastReceiver()
+                    val filter = IntentFilter()
+                    filter.addAction(Intent.ACTION_USER_PRESENT)
+                    phoneUnlockBR?.apply {
+                        context?.registerReceiver(this as BroadcastReceiver, filter)
+                    }
+                    showUnlockScreenNotification()
+                } else {
+                    val myLock: KeyguardManager.KeyguardLock = km.newKeyguardLock("AnswerCall")
+                    myLock?.disableKeyguard()
+                }
+            }
+        } else {
+            acceptIncomingVoIP()
+        }
     }
-  }
+
+    fun acceptIncomingVoIP() {
+        Log.d("IC", "acceptIncomingVoIP")
+        IncomingCallHelper.handleActionCall(applicationContext, intent, VOIP_ACCEPT)
+    }
+
+    private fun declineIncomingVoIP() {
+        Log.d("IC", "declineIncomingVoIP")
+        IncomingCallHelper.handleActionCall(applicationContext, intent, VOIP_DECLINE)
+    }
+
+    @SuppressLint("MissingPermission")
+    private fun showUnlockScreenNotification() {
+        val notificationBuilder = NotificationCompat.Builder(this, PushConstants.DEFAULT_CHANNEL_ID)
+            .setSmallIcon(resources.getIdentifier("pushicon", "drawable", packageName))
+            .setContentTitle("Ongoing call with $caller")
+            .setContentText("Please unlock your device to continue")
+            .setPriority(NotificationCompat.PRIORITY_HIGH)
+            .setCategory(NotificationCompat.CATEGORY_MESSAGE)
+            .setAutoCancel(false)
+            .setOngoing(true)
+            .setStyle(NotificationCompat.BigTextStyle())
+            .setSound(null)
+        val ongoingCallNotification = notificationBuilder.build()
+        val notificationManager = NotificationManagerCompat.from(this.applicationContext)
+        // Display notification
+        if (!isPostNotificationsGranted()) {
+            requestPostNotifications()
+        } else {
+            notificationManager.notify(NOTIFICATION_MESSAGE_ID, ongoingCallNotification)
+        }
+    }
+
+    private fun isPostNotificationsGranted(): Boolean {
+        return ActivityCompat.checkSelfPermission(
+            this,
+            Manifest.permission.POST_NOTIFICATIONS
+        ) == PackageManager.PERMISSION_GRANTED
+    }
+
+    private fun requestPostNotifications() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            requestPermissions(
+                arrayOf(Manifest.permission.POST_NOTIFICATIONS),
+                POST_NOTIFICATIONS_REQUEST_CODE
+            )
+        }
+    }
+
+    override fun onRequestPermissionsResult(
+        requestCode: Int,
+        permissions: Array<out String>,
+        grantResults: IntArray
+    ) {
+        if (requestCode == POST_NOTIFICATIONS_REQUEST_CODE &&
+            grantResults.getOrNull(0) == PackageManager.PERMISSION_GRANTED
+        ) {
+            showUnlockScreenNotification()
+        }
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        Log.d("", "IncomingCallActivity.onCreate()")
+        instance = null
+    }
+
+    class PhoneUnlockBroadcastReceiver : BroadcastReceiver() {
+        override fun onReceive(context: Context, intent: Intent) {
+            if (intent.action.equals(Intent.ACTION_USER_PRESENT)) {
+                dismissUnlockScreenNotification(context.applicationContext)
+            }
+        }
+    }
+
+    companion object {
+
+        const val VOIP_CONNECTED = "connected"
+        const val VOIP_ACCEPT = "pickup"
+        const val VOIP_DECLINE = "declined_callee"
+        private const val NOTIFICATION_MESSAGE_ID = 1337
+
+        var instance: IncomingCallActivity? = null
+
+        var phoneUnlockBR: PhoneUnlockBroadcastReceiver? = null
+        fun dismissUnlockScreenNotification(applicationContext: Context) {
+            NotificationManagerCompat.from(applicationContext).cancel(NOTIFICATION_MESSAGE_ID)
+            if (phoneUnlockBR != null) {
+                applicationContext.unregisterReceiver(phoneUnlockBR)
+                phoneUnlockBR = null
+            }
+        }
+    }
 }
