+++ src/android/com/adobe/phonegap/push/OnNotificationReceiverActivity.kt
@@ -1,18 +1,20 @@
 package com.adobe.phonegap.push
 
 import android.app.Activity
+import android.content.Context
+import android.content.Intent
+import android.os.Bundle
+import android.util.Log
 
 class OnNotificationReceiverActivity : Activity() {
-  @Override
-  protected fun onCreate(savedInstanceState: Bundle?) {
+  override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
     Log.d(LOG_TAG, "OnNotificationReceiverActivity.onCreate()")
     handleNotification(this, getIntent())
     finish()
   }
 
-  @Override
-  protected fun onNewIntent(intent: Intent) {
+  override fun onNewIntent(intent: Intent) {
     super.onNewIntent(intent)
     Log.d(LOG_TAG, "OnNotificationReceiverActivity.onNewIntent()")
     handleNotification(this, intent)
@@ -23,18 +25,20 @@ class OnNotificationReceiverActivity : Activity() {
     private const val LOG_TAG = "Push_OnNotificationReceiverActivity"
     private fun handleNotification(context: Context, intent: Intent) {
       try {
-        val pm: PackageManager = context.getPackageManager()
-        val launchIntent: Intent = pm.getLaunchIntentForPackage(context.getPackageName())
-        launchIntent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP or Intent.FLAG_ACTIVITY_NEW_TASK)
-        val data: Bundle = intent.getExtras()
-        if (!data.containsKey("messageType")) data.putString("messageType", "notification")
-        data.putString("tap", if (PushPlugin.isInBackground()) "background" else "foreground")
+        val pm = context.packageManager
+        val launchIntent = pm.getLaunchIntentForPackage(context.getPackageName())
+        launchIntent?.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP or Intent.FLAG_ACTIVITY_NEW_TASK)
+        val data = intent.extras
+        if (data?.containsKey("messageType") == false ) { data?.putString("messageType", "notification") }
+        data?.putString("tap", if (PushPlugin.isInBackground) "background" else "foreground")
         Log.d(LOG_TAG, "OnNotificationReceiverActivity.handleNotification(): " + data.toString())
         PushPlugin.sendExtras(data)
-        launchIntent.putExtras(data)
+        data?.apply {
+          launchIntent?.putExtras(data)
+        }
         context.startActivity(launchIntent)
       } catch (e: Exception) {
-        Log.e(LOG_TAG, e.getLocalizedMessage(), e)
+        Log.e(LOG_TAG, e.localizedMessage, e)
       }
     }
   }
